arm: imx: Make i.MX7D select CFG_MX7D
Currently when we make an i.MX7 based OPTEE image the value selected is
CFG_MX7 - however for different versions of the i.MX7 we will probably need
to differentiate between the different types of i.MX7. This patch is a
precursor to make a i.MX7Solo based platform addition. While we may be able
to get away with using the CFG_MX7 define and assumptions as they are, it
is good practice to separate processor types as later software may actually
care about the difference at compile time.

This patch therefore does a simple conversion g/CFG_MX7/s//CFG_MX7D/g a
subsequent patch will add CFG_MX7S with an i.MXSolo board to go with it.

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>